Revert "Support framed FSK Manchester decoding compatible with Si4320/RFM01"
... and also "Add heuristic for detecting RFM01-framed FSK"

This reverts commits 3470b458 and 712f97e4 to remove the newly-added
pulse_demod_manchester_framed() functionality.

It turns out this isn't done by the hardware; it's application-specific.
The watchman code has been converted to use FSK_PULSE_PCM and handle the
framing and the Manchester decoding in the bitbuffer domain instead, and
this code can die since it wasn't as general-purpose as originally hoped.

See issue #224.