Pass 0 into local_time_str() for current time instead of using time()
When using input from a file, we're going to want to print the offset in
the file at which the signal was found, not the current time. So it's OK
to use local_time_str() with an explicit time when it's actually data from
within the decoded packet, but for the current time just use zero
consistently.