Add temperature output in °C to Watchman decode
We're fairly convinced we have a reasonable estimate of the what the
temperature values mean.

See https://hackaday.io/project/1548-rocket-receiver-teardown for all
the gory details.

Keep outputting the raw value too, just in case we change it later; a
consumer of this data which thinks it knows better can still reconvert
for itself.