Revert "Update pulse_detect.c" This reverts commit ef163f9d575a5322df46c8c87eb9565802d4f7af.
This commit is faulty and generates segmentation faults. If
hist_gaps.bins_count is 2 the index will go negative. The code assumes
there will allways be both short and long gaps (messages with multiple rows),
and will fail when it is a pure PPM signal.
The precense of too-short gaps will also fool the PPM demodulator and insert
spurious bits. If possible it should be fixed by making the pulse-detector
more resilient to too-short gaps (test signals needed).