Added fineoffset_wh31.c to process 915 MHz Fine Offset WH31 sensors. This is a preliminary file. More work is needed to fully decode the device.
Be kind to programmers week? Anyway added a warning if MAX_PROTOCOLS is potentially too small with a suggested minimum size as soon as num_r_devices is calculated. In debug mode this results in a hard exit. I moved MAX_PROTOCOLS into rtl_433_devices.h as a reminder to programmers to update it.