Add checksum for TFA Drop 30.3233.01
This commit introduces a message integrity check by computing the
checksum for incoming messages and comparing with the checksum
provided as part of the incoming message.

The details of the algorithms were kindly provided by @zuckschwerdt
in https://github.com/merbanan/rtl_433/issues/1240.

This commit also enables the decoder by default, as it can now
validate checksums and avoid false positives.

Finally, the code is now formatted to a line length of 72 characters.