cosmetics: remove unused variables and function params or comment them
This results in a lot less noise when -Wunused and -Wunused-parameter
are enabled, and cleans up the code base from old prototyping code.

Some unused variables are commented out because they appear to be
values that might be of value someday, but there were a lot of values
that could just be removed because they did not appear to be of any
value.