Fix acurite-985 false positives with all zeros
Once in a while, data of all zeros of the correct length
would pass all the checks including the CRC but that
clearly were not valid.  acurite_985_decode() will now
bail if the data bytes are all zeros.

This was seen as a refrigerator temperature going from the 30s
to 0 and then back to the 30s over 3 consecutive samples.
The id was also going from non-zero, to 0, and back to the original value.