fineoffset: Add explanation of supposed PM10 value
After discussion in https://github.com/merbanan/rtl_433/issues/1640
this commit adjusts the comment about the WH0290 particulate sensor,
mostly surrounding the nature of the field labeled PM10.

The measurement of the sensor serial port was done by @peteakalad, and
the determination of the pm2.5/pm10 conversion was done by @gdt with
input from @peteakalad.

(This commit does not change variable names or API-visible text.)