Move tower and leak detector decode out-of-line
The generic AcuRite decode function was getting out of hand. After initial pull request feedback, it was determined that this should be handled in the short term by moving the decode for this stuff out-of-line.
This makes the code more readable for now, but results in some duplication. The eventual solution is a refactor of the whole family of AcuRite decoders.