cosmetics: acurite_01185m: rename variable
Renamed variable 'chk' to 'sum'. The variable is used for
calculating the checksum, however the checksum is really
'chk & 0xff', while chk contained a sum of bytes.

This was fine so far, but now we use 'chk' (now: 'sum') to
check for very likely false positives (all bits 0) and there
it is quite different whether you check for the sum == 0 or
just the real checksum == 0, which would only be the lowest
eight bits.