examples: Improve README by describing scripts (#2976)
Expand description of the directory, making it clearer that some
programs have production use.

Declare that python scripts are python3.  It's clear that nobody is
trying to use them with python 2.7, and if they did, we would tell
them not to.

List the scripts with a one-line description, sorting them into
production (none yet), generally usable, and true examples.  Leave
some uncategorized.