docker: Add official local docker support
Lets add our own docker container and use github actions to build and
push it.

For now, only the alpine container is built and pushed, but we can very
easily add a matrix entry for a debian container if there's enough
demand.

Best to reduce the amount of configurations we offer, as it also means
more that can break.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>