Add support for Hornbach MSRC-SAL Awning remote
This commit adds support for the Hornbach Awning 433.92Mhz remote.

Some uncertainty remains on the Remote controls Unique ID, as it was
only possible to test with a single remote.

As this is my first time analyzing RF data, I'm also not certain on the
timings. Triq PDV was used (and some spreadsheet weighted averages) to find
these values. The sync, reset and gap timings are the most uncertain, as
well as possible tolerances etc.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>