Fix visonic decode even if good package is not the first in buffer
I often receive data where the first package is corrupted/too short,
but the following 2 to 5 are correct. Therefore make the code work
when the first good packet is not the very first in the buffer.
Also allow to decode when only 2 repeated packets are received.
Since there is a checksum in the packet, it might also be enough
to use the first good packet, but let's stick to 2 for now.