Add support for La Crosse View LTV-TH2i Thermo/Hygro Sensor
Much like its predecessor the LTV-TH2, the CRC on some of the packets
transmitted by the LTV-TH2i will be calculated with a fixed nonzero
initial value. Unlike the LTV-TH2, the TH2i uses init=0xb2. The
lacrosse_th3 decoder is still able to decode the TH2i packets with
init=0x00 CRC values, but it discards the majority of the reports as
corrupt. Add a check for init=0x2b CRC to the decoder to improve its
support for decoding packets from LTV-TH2i sensors.