Change LaCrosse LTV-TH decoder to output battery_ok, retransmit
Experimentation with a La Crosse View LTV-TH2i sensor has revealed the
semantics of two of the flags. One signals a low battery, and the other
signals that the packet is a repeat broadcast triggered by a button
press. (The La Crosse V41CA display audibly beeps upon reception of a
packet with this flag set.) Break these flags out into their own output
fields and squelch the display of the remaining unknown flags unless one
of them is set.